feat: Add agents handler with CRUD operations
